LayoutTests/platform/chromium-mac-leopard/platform/chromium/media/video-scales-in-media-document-expected.txt

Test that video media documents scale to fit undersized containers.


EXPECTED (200 == '200') OK
EXPECTED (164 == '163'), OBSERVED '164' FAIL
END OF TEST
